port split.rb to the main tool
