+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM golang:1.13-alpine as builder
+FROM golang:1.13.3-alpine as builder
 
 ENV GOPROXY="https://proxy.golang.org"
 ENV GO111MODULE="on"
@@ -8,9 +8,9 @@ RUN apk add --no-cache git
 WORKDIR /go/src/github.com/icco/postmortems
 COPY . .
 
-RUN cd tool && go build -o /go/bin/tool .
-RUN /go/bin/tool -action=validate
-RUN /go/bin/tool -action=generate
+RUN cd tool && go build -o /go/bin/pm .
+RUN /go/bin/pm -action=validate
+RUN /go/bin/pm -action=generate
 
 FROM halverneus/static-file-server:latest
 COPY --from=builder /go/src/github.com/icco/postmortems/output/ /web/
