--- tmp/Gemfile
@@ -1,5 +1,0 @@
-source "https://rubygems.org"
-
-ruby "2.6.2"
-
-gem "front_matter_parser"
