+++ tool/.gitignore
@@ -1,1 +1,1 @@
-tool
+pm
